+++ Assets/NativeStringCollections/Scripts/csFastFloat/FastFloatParser.cs
@@ -15,13 +15,7 @@ namespace NativeStringCollections.Impl.csFastFloat
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         internal static float Exact_power_of_ten(long power)
         {
-#if NET5_0
-      Debug.Assert(power < CalculationConstants.powers_of_ten_float.Length);
-      ref float tableRef = ref MemoryMarshal.GetArrayDataReference(CalculationConstants.powers_of_ten_float);
-      return Unsafe.Add(ref tableRef, (nint)power);
-#else
             return CalculationConstants.powers_of_ten_float[power];
-#endif
         }
 
         /// <summary>
@@ -68,7 +62,7 @@ namespace NativeStringCollections.Impl.csFastFloat
         /// <param name="decimal_separator">decimal separator to be used</param>
         /// <returns>bool : true is sucessfuly parsed</returns>
         public static bool TryParseFloat<T>(T source, out float result, UInt16 decimal_separator = UTF16CodeSet.code_dot)
-            where T : IParseExt
+            where T : IJaggedArraySliceBase<Char16>
         {
             Char16* pStart = (Char16*)source.GetUnsafePtr();
             if (!TryParseNumber(pStart, pStart + (uint)source.Length, out _, out result, decimal_separator))
@@ -134,13 +128,6 @@ namespace NativeStringCollections.Impl.csFastFloat
 
         // UTF-16 inputs
         internal static AdjustedMantissa ParseLongMantissa(Char16* first, Char16* last, Char16 decimal_separator)
-        {
-            DecimalInfo d = DecimalInfo.parse_decimal(first, last, decimal_separator);
-            return ComputeFloat(d);
-        }
-
-        // UTF-8/ASCII inputs
-        internal static AdjustedMantissa ParseLongMantissa(byte* first, byte* last, byte decimal_separator)
         {
             DecimalInfo d = DecimalInfo.parse_decimal(first, last, decimal_separator);
             return ComputeFloat(d);