+++ Assets/NativeStringCollections/Scripts/csFastFloat/Structures/DecimalInfo.cs
@@ -44,13 +44,7 @@ namespace NativeStringCollections.Impl.csFastFloat.Constants
 
         private static ushort get_number_of_digits_decimal_left_shift(int shift)
         {
-#if NET5_0
-      Debug.Assert(shift < number_of_digits_decimal_left_shift_table.Length);
-      ref ushort tableRef = ref MemoryMarshal.GetArrayDataReference(number_of_digits_decimal_left_shift_table);
-      return Unsafe.Add(ref tableRef, (IntPtr)(uint)shift);
-#else
             return number_of_digits_decimal_left_shift_table[shift];
-#endif
         }
 
         private static readonly ushort[] number_of_digits_decimal_left_shift_table = {
@@ -412,107 +406,6 @@ namespace NativeStringCollections.Impl.csFastFloat.Constants
                     if (exp_number < 0x10000)
                     {
                         exp_number = 10 * exp_number + digit;
-                    }
-                    ++p;
-                }
-                answer.decimal_point += (neg_exp ? -exp_number : exp_number);
-            }
-            // In very rare cases, we may have fewer than 19 digits, we want to be able to reliably
-            // assume that all digits up to max_digit_without_overflow have been initialized.
-            for (uint i = answer.num_digits; i < CalculationConstants.max_digit_without_overflow; i++) { answer.digits[i] = 0; }
-
-            return answer;
-        }
-
-        // UTF-8/ASCII inputs
-        unsafe internal static DecimalInfo parse_decimal(byte* p, byte* pend, byte decimal_separator)
-        {
-            DecimalInfo answer = new DecimalInfo() { negative = (*p == '-') };
-
-            if ((*p == '-') || (*p == '+'))
-            {
-                ++p;
-            }
-            // skip leading zeroes
-            while ((p != pend) && (*p == '0'))
-            {
-                ++p;
-            }
-            while ((p != pend) && Utils.is_integer(*p, out uint digit))
-            {
-                if (answer.num_digits < CalculationConstants.max_digits)
-                {
-                    answer.digits[answer.num_digits] = (byte)digit;
-                }
-                answer.num_digits++;
-                ++p;
-            }
-            if ((p != pend) && (*p == decimal_separator))
-            {
-                ++p;
-                byte* first_after_period = p;
-                // if we have not yet encountered a zero, we have to skip it as well
-                if (answer.num_digits == 0)
-                {
-                    // skip zeros
-                    while ((p != pend) && (*p == '0'))
-                    {
-                        ++p;
-                    }
-                }
-                while ((p != pend) && Utils.is_integer(*p, out uint digit))
-                {
-                    if (answer.num_digits < CalculationConstants.max_digits)
-                    {
-                        answer.digits[answer.num_digits] = (byte)digit;
-                    }
-                    answer.num_digits++;
-                    ++p;
-                }
-                answer.decimal_point = (int)(first_after_period - p);
-            }
-            // We want num_digits to be the number of significant digits, excluding
-            // leading *and* trailing zeros! Otherwise the truncated flag later is
-            // going to be misleading.
-            if (answer.num_digits > 0)
-            {
-                // We potentially need the answer.num_digits > 0 guard because we
-                // prune leading zeros. So with answer.num_digits > 0, we know that
-                // we have at least one non-zero digit.
-                byte* preverse = p - 1;
-                int trailing_zeros = 0;
-                while ((*preverse == '0') || (*preverse == decimal_separator))
-                {
-                    if (*preverse == '0') { trailing_zeros++; };
-                    --preverse;
-                }
-                answer.decimal_point += (int)(answer.num_digits);
-                answer.num_digits -= (uint)(trailing_zeros);
-            }
-            if (answer.num_digits > CalculationConstants.max_digits)
-            {
-                answer.truncated = true;
-                answer.num_digits = CalculationConstants.max_digits;
-            }
-            if ((p != pend) && (('e' == *p) || ('E' == *p)))
-            {
-                ++p;
-                bool neg_exp = false;
-                if ((p != pend) && ('-' == *p))
-                {
-                    neg_exp = true;
-                    ++p;
-                }
-                else if ((p != pend) && ('+' == *p))
-                {
-                    ++p;
-                }
-                int exp_number = 0; // exponential part
-                while ((p != pend) && Utils.is_integer(*p, out uint digit))
-                {
-                    if (exp_number < 0x10000)
-                    {
-                        exp_number = 10 * exp_number + (int)digit;
                     }
                     ++p;
                 }